feat: restructure failure classification and RBS handling
- Moved FailureClassification module to Karya::Internal for better encapsulation.
- Updated references in job attributes and queue store to use the new internal path.
- Removed outdated failure classification files and signatures.
- Enhanced documentation to clarify RBS usage and internal helpers.
- Added tests for the new internal FailureClassification module.
